dashboard/requirements.txt
streamlit
pandas
psycopg2-binary